[user][feat] 유저 프로필 디테일 (카테고리, 지역) 선택 페이지 생성
