+++ src/main/java/com/bootproj/pmcweb/Controller/AccountController.java
@@ -8,20 +8,25 @@ import com.bootproj.pmcweb.Common.Exception.NoMatchingAcountException;
 import com.bootproj.pmcweb.Common.Exception.PasswordNotMatchException;
 import com.bootproj.pmcweb.Common.Exception.SendEmailException;
 import com.bootproj.pmcweb.Common.Header;
+import com.bootproj.pmcweb.Domain.Study;
 import com.bootproj.pmcweb.Service.AccountSecurityService;
 import com.bootproj.pmcweb.Service.AccountService;
 import com.bootproj.pmcweb.Service.AttachmentService;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.security.core.userdetails.User;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.multipart.MultipartFile;
 import org.springframework.web.servlet.ModelAndView;
 
 import javax.validation.Valid;
+import java.util.HashMap;
 import java.util.Map;
 import java.util.Optional;
 
@@ -37,64 +42,6 @@ public class AccountController {
 
     @Autowired
     private AttachmentService attachmentService;
-
-    /**
-     * Security signup login logout
-     * made by kym
-     */
-
-    // 회원가입
-    @GetMapping("/user/signup")
-    public String getSignup(){
-        return "user/register";
-    }
-
-    // 로그인
-    @GetMapping("/user/login")
-    public String getLogin(){
-        return "user/login";
-    }
-
-    // 로그인 실패
-    @PostMapping("/user/loginFail")
-    public String getLoginFail(){
-        return "user/loginFail";
-    }
-
-    // 로그아웃
-    @GetMapping("/user/logout")
-    public String getLogout(){
-        return "user/login";
-    }
-
-    // 프로필 화면
-    @GetMapping("/user/profile")
-    public ModelAndView getProfile(@AuthenticationPrincipal User user) throws NoMatchingAcountException {
-        Account account = accountService.getUserByEmail(user.getUsername());
-        ModelAndView mv = new ModelAndView("/user/profile");
-        mv.addObject("loginUser", account.getName());
-        mv.addObject("loginUserEmail", account.getEmail());
-        mv.addObject("loginUserId", account.getId());
-        Optional<Attachment> attachment = attachmentService.getProfile(user.getUsername());
-        attachment.ifPresentOrElse(
-                (att) -> {
-                    mv.addObject("profileImagePath", "/img/profile/" + account.getId() + "/" + att.getName());
-                    log.info("/img/profile/" + account.getId() + "/" + att.getName());
-                },
-                () -> {
-                    mv.addObject("profileImagePath", "/img/moim.jpg");
-                    log.info("/img/moim.jpg");
-                }
-        );
-
-        return mv;
-    }
-
-    // 비밀번호 변경 화면
-    @GetMapping("/user/changePassword")
-    public String changePassword() {
-        return "user/changePassword";
-    }
 
     /**
      * REST API
@@ -121,7 +68,6 @@ public class AccountController {
             log.info(e.getMessage());
             return "redirect:/user/changePassword";
         }
-
         return "redirect:/user/profile";
     }
 
@@ -149,5 +95,11 @@ public class AccountController {
         Account changedUser = accountService.signUpConfirm(authKey, email);
         accountService.signUpConfirm(authKey, email);
         return "redirect:/user/login";
+    }
+
+    @PostMapping("/user/updateProfile")
+    public ResponseEntity<Header> updateProfile(@RequestBody User user){
+
+        return new ResponseEntity(Header.OK(), HttpStatus.OK);
     }
 }
